Refactored folder structure to include DTO
