+++ src/main/java/SMS/schoolmanagementsystem/controllers/GradesController.java
@@ -1,11 +1,11 @@
 package SMS.schoolmanagementsystem.controllers;
 
 import SMS.schoolmanagementsystem.models.Grade;
-import SMS.schoolmanagementsystem.models.OverallStudentGrade;
+
+import SMS.schoolmanagementsystem.models.dto.OverallStudentGradeDto;
 import SMS.schoolmanagementsystem.repositories.GradesRepository;
 import SMS.schoolmanagementsystem.services.GradeService;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -31,7 +31,7 @@ public class GradesController {
     }
 
     @GetMapping("/student/module-grade")
-    OverallStudentGrade getStudentModuleGrade(@RequestParam(required = true) int userId, @RequestParam(required = true) int moduleId) {
+    OverallStudentGradeDto getStudentModuleGrade(@RequestParam(required = true) int userId, @RequestParam(required = true) int moduleId) {
         List<Grade> response = gradesRepository.getStudentModuleGrade(userId, moduleId);
         return GradeService.getOverallGrade(response);
     }
